doneeeeeeee

+++ api/urls/urls.controllers.js
@@ -4,15 +4,15 @@ const User = require("../../models/User");
 
 const baseUrl = "http:localhost:8000/urls";
 
-exports.shorten = async (req, res) => {
+exports.shorten = async (req, res, next) => {
   // create url code
   const urlCode = shortid.generate();
   try {
     req.body.shortUrl = `${baseUrl}/${urlCode}`;
     req.body.urlCode = urlCode;
-    req.body.userId = req.params.userId;
+    req.body.userId = req.User.userId;
     const newUrl = await Url.create(req.body);
-    await User.findByIdAndUpdate(req.params.userId, {
+    await User.findByIdAndUpdate(req.User.userId, {
       $push: { urls: newUrl._id },
     });
     res.json(newUrl);
@@ -21,7 +21,7 @@ exports.shorten = async (req, res) => {
   }
 };
 
-exports.redirect = async (req, res) => {
+exports.redirect = async (req, res, next) => {
   try {
     const url = await Url.findOne({ urlCode: req.params.code });
     if (url) {
@@ -34,12 +34,16 @@ exports.redirect = async (req, res) => {
   }
 };
 
-exports.deleteUrl = async (req, res) => {
+exports.deleteUrl = async (req, res, next) => {
   try {
     const url = await Url.findOne({ urlCode: req.params.code });
     if (url) {
-      await Url.findByIdAndDelete(url._id);
-      return res.status(201).json("Deleted");
+      if (url.userId.equals(req.user._id)) {
+        await Url.findByIdAndDelete(url._id);
+        return res.status(201).json("Deleted");
+      } else {
+        return res.status(401).json("you dont have permission");
+      }
     } else {
       return res.status(404).json("No URL Found");
     }
